requirements.txt
streamlit
langchain
google-generativeai
faiss-cpu
openai
